+++ client/web/src/user/settings/research/ProductResearch.test.tsx
@@ -1,18 +1,17 @@
 import React from 'react'
 import { render, RenderResult } from '@testing-library/react'
-import * as sinon from 'sinon'
-import { ProductResearchPage } from './ProductResearch'
-import { AuthenticatedUser } from '../../../auth'
-import { TelemetryService } from '../../../../../shared/src/telemetry/telemetryService'
+import { ProductResearchArea } from './ProductResearch'
+import { NOOP_TELEMETRY_SERVICE } from '../../../../../shared/src/telemetry/telemetryService'
 
-describe('ProductResearchPage', () => {
+describe('ProductResearchArea', () => {
     let queries: RenderResult
-    const mockTelemetryService = { logViewEvent: sinon.spy(), log: sinon.spy() } as TelemetryService
-    const mockAuthenticatedUser = { email: 'test@sourcegraph.com' } as AuthenticatedUser
 
     beforeEach(() => {
         queries = render(
-            <ProductResearchPage telemetryService={mockTelemetryService} authenticatedUser={mockAuthenticatedUser} />
+            <ProductResearchArea
+                telemetryService={NOOP_TELEMETRY_SERVICE}
+                authenticatedUser={{ email: 'test@sourcegraph.com' }}
+            />
         )
     })
 
@@ -22,7 +21,7 @@ describe('ProductResearchPage', () => {
 
     test('renders sign up now link correctly', () => {
         expect(queries.getByText('Sign up now').closest('a')?.href).toMatchInlineSnapshot(
-            '"https://share.hsforms.com/1tkScUc65Tm-Yu98zUZcLGw1n7ku?email=test@sourcegraph.com"'
+            '"https://share.hsforms.com/1tkScUc65Tm-Yu98zUZcLGw1n7ku?email=test%40sourcegraph.com"'
         )
     })
 })