+++ client/web/src/user/settings/sidebaritems.ts
@@ -50,7 +50,7 @@ export const userSettingsSideBarItems: UserSettingsSidebarItems = {
         {
             label: 'Product research',
             to: '/product-research',
-            condition: () => window.context.productResearchPageEnabled,
+            condition: () => window.context.productResearchAreaEnabled,
         },
     ],
 }